+++ canon-docs/src/components/CustomTheme/customTheme.tsx
@@ -17,7 +17,7 @@ const defaultTheme = `:root {
 const myTheme = createTheme({
   theme: 'light',
   settings: {
-    background: 'var(--canon-bg-elevated)',
+    background: 'var(--canon-bg-surface-1)',
     backgroundImage: '',
     foreground: '#6182B8',
     caret: '#5d00ff',
@@ -125,7 +125,7 @@ export const CustomTheme = () => {
                 className={styles.buttonClose}
                 onClick={() => setOpen(!open)}
               >
-                <Icon name={open ? 'chevronDown' : 'chevronUp'} />
+                <Icon name={open ? 'chevron-down' : 'chevron-up'} />
               </button>
             </div>
           </div>