+++ packages/canon/src/components/Heading/Heading.stories.tsx
@@ -18,7 +18,6 @@ import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { Heading } from './Heading';
 import { Flex } from '../Flex';
-import { Text } from '../Text';
 
 const meta = {
   title: 'Components/Heading',
@@ -42,7 +41,7 @@ export const Title1: Story = {
 
 export const AllVariants: Story = {
   render: () => (
-    <Flex>
+    <Flex direction="column" gap="4">
       <Heading variant="display">Display</Heading>
       <Heading variant="title1">Title 1</Heading>
       <Heading variant="title2">Title 2</Heading>
@@ -68,10 +67,22 @@ export const CustomTag: Story = {
   },
 };
 
+export const WrappedInLink: Story = {
+  args: {
+    ...Default.args,
+  },
+  decorators: [
+    Story => (
+      <a href="/">
+        <Story />
+      </a>
+    ),
+  ],
+};
+
 export const Playground: Story = {
   render: () => (
-    <Flex>
-      <Text>All variants</Text>
+    <Flex direction="column" gap="4">
       <Heading variant="display">Display</Heading>
       <Heading variant="title1">Title 1</Heading>
       <Heading variant="title2">Title 2</Heading>