+++ packages/cli/src/alpha.ts
@@ -18,7 +18,8 @@ import { CliInitializer } from './wiring/CliInitializer';
 import chalk from 'chalk';
 
 (async () => {
-  console.warn(
+  // Logging as stdout to prevent failures from stderr during api-report generation.
+  console.log(
     chalk.yellow(
       'THIS ENTRYPOINT IS IN ALPHA AND MAY CHANGE IN THE FUTURE - DO NOT USE THIS FOR NORMAL DEVELOPMENT',
     ),
@@ -29,5 +30,6 @@ import chalk from 'chalk';
   initializer.add(import('./modules/build/alpha'));
   initializer.add(import('./modules/migrate/alpha'));
   initializer.add(import('./modules/test/alpha'));
+  initializer.add(import('./modules/lint/alpha'));
   await initializer.run();
 })();