+++ packages/cli/src/lib/runner/runBackend.ts
@@ -32,14 +32,16 @@ const loaderArgs = [
 ];
 
 export type RunBackendOptions = {
+  /** The directory to run the backend process in, defaults to cwd */
+  targetDir?: string;
   /** relative entry point path without extension, e.g. 'src/index' */
   entry: string;
   /** Whether to forward the --inspect flag to the node process */
-  inspectEnabled: boolean;
+  inspectEnabled?: boolean | string;
   /** Whether to forward the --inspect-brk flag to the node process */
-  inspectBrkEnabled: boolean;
+  inspectBrkEnabled?: boolean | string;
   /** Additional module to require via the --require flag to the node process */
-  require?: string;
+  require?: string | string[];
   /** An external linked workspace to override module resolution towards */
   linkedWorkspace?: string;
 };
@@ -107,7 +109,10 @@ export async function runBackend(options: RunBackendOptions) {
       optionArgs.push(inspect);
     }
     if (options.require) {
-      optionArgs.push(`--require=${options.require}`);
+      const requires = [options.require].flat();
+      for (const r of requires) {
+        optionArgs.push(`--require=${r}`);
+      }
     }
 
     const userArgs = process.argv
@@ -119,6 +124,7 @@ export async function runBackend(options: RunBackendOptions) {
       [...loaderArgs, ...optionArgs, options.entry, ...userArgs],
       {
         stdio: ['ignore', 'inherit', 'inherit', 'ipc'],
+        cwd: options.targetDir,
         env: {
           ...process.env,
           BACKSTAGE_CLI_LINKED_WORKSPACE: options.linkedWorkspace,