+++ packages/integration/src/gerrit/core.ts
@@ -292,6 +292,7 @@ export function getAuthenticationPrefix(
  * be used.
  *
  * @param config - A Gerrit provider config.
+ * @public
  */
 export function getGitilesAuthenticationUrl(
   config: GerritIntegrationConfig,