+++ plugins/search-backend-module-techdocs/src/collators/index.ts
@@ -14,10 +14,6 @@
  * limitations under the License.
  */
 
-export { DefaultTechDocsCollatorFactory } from './DefaultTechDocsCollatorFactory';
-
-export type { TechDocsCollatorFactoryOptions } from './DefaultTechDocsCollatorFactory';
-
 export { defaultTechDocsCollatorEntityTransformer } from './defaultTechDocsCollatorEntityTransformer';
 
 export type { TechDocsCollatorEntityTransformer } from './TechDocsCollatorEntityTransformer';