+++ packages/app-next-example-plugin/CHANGELOG.md
@@ -1,5 +1,29 @@
 # app-next-example-plugin
 
+## 0.0.22-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.17.1-next.0
+  - @backstage/frontend-plugin-api@0.10.1-next.0
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.17.0
+  - @backstage/frontend-plugin-api@0.10.0
+
+## 0.0.21-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.10.0-next.2
+  - @backstage/core-components@0.16.5-next.1
+
 ## 0.0.21-next.1
 
 ### Patch Changes