+++ plugins/kubernetes-common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-kubernetes-common
 
+## 0.9.4
+
+### Patch Changes
+
+- d517d13: Export `backstage.io/kubernetes-id` and `backstage.io/kubernetes-label-selector` annotations as constants
+- Updated dependencies
+  - @backstage/catalog-model@1.7.3
+  - @backstage/types@1.2.1
+  - @backstage/plugin-permission-common@0.8.4
+
 ## 0.9.4-next.0
 
 ### Patch Changes