+++ plugins/scaffolder-backend-module-yeoman/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @backstage/plugin-scaffolder-backend-module-yeoman
 
+## 0.4.9-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.8.1-next.1
+  - @backstage/backend-plugin-api@1.2.1
+  - @backstage/types@1.2.1
+  - @backstage/plugin-scaffolder-node-test-utils@0.2.1-next.1
+
+## 0.4.9-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.8.1-next.0
+  - @backstage/backend-plugin-api@1.2.1
+  - @backstage/types@1.2.1
+  - @backstage/plugin-scaffolder-node-test-utils@0.2.1-next.0
+
+## 0.4.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.8.0
+  - @backstage/plugin-scaffolder-node-test-utils@0.2.0
+  - @backstage/backend-plugin-api@1.2.1
+  - @backstage/types@1.2.1
+
+## 0.4.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.8.0-next.2
+  - @backstage/plugin-scaffolder-node-test-utils@0.2.0-next.2
+  - @backstage/backend-plugin-api@1.2.1-next.1
+  - @backstage/types@1.2.1
+
 ## 0.4.8-next.1
 
 ### Patch Changes