+++ plugins/techdocs-react/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @backstage/plugin-techdocs-react
 
+## 1.2.16-next.0
+
+### Patch Changes
+
+- 0e9f7fe: Fix catalog entity docs page not loading due to multiple addons data attachment in the New Frontend System.
+- Updated dependencies
+  - @backstage/core-components@0.17.1-next.0
+  - @backstage/frontend-plugin-api@0.10.1-next.0
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/core-plugin-api@1.10.5
+  - @backstage/version-bridge@1.0.11
+
+## 1.2.15
+
+### Patch Changes
+
+- b5a8208: Added `TechDocsAddonsBlueprint` extension to allow adding of techdocs addons.
+- Updated dependencies
+  - @backstage/core-components@0.17.0
+  - @backstage/core-plugin-api@1.10.5
+  - @backstage/frontend-plugin-api@0.10.0
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/version-bridge@1.0.11
+
+## 1.2.15-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.10.0-next.2
+  - @backstage/core-components@0.16.5-next.1
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/core-plugin-api@1.10.4
+  - @backstage/version-bridge@1.0.11
+
 ## 1.2.15-next.1
 
 ### Patch Changes